requirements.txt
appdirs==1.4.3
#aiofiles==0.8.0
#aiobotocore==2.3.4
#aiohttp==3.8.1
asgiref==3.3.0
boto3==1.24.23
CacheControl==0.12.6
certifi==2019.11.28
celery==5.0.5
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.1.3
django-debug-toolbar==3.1.1
django-filter==21.1
django-storages==1.9.1
django-push-notifications==3.0.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
django-rest-swagger==2.2.0
dnspython==2.2.1
drf-yasg==1.20.0
drf-spectacular==0.22.1
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
mysqlclient==2.1.0
packaging==20.3
pep517==0.8.2
pillow==9.2.0
progress==1.5
pyparsing==2.4.6
python-decouple==3.3
python-dotenv==0.20.0
pytoml==0.1.21
pytz==2020.4
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.4.1
urllib3==1.25.8
webencodings==0.5.1